moving stateProofMessage from crypto/compcert to data/stateproof
