+++ data/stateproof/message.go
@@ -1,4 +1,4 @@
-package compactcert
+package stateproof
 
 import (
 	"github.com/algorand/go-algorand/crypto"
@@ -6,19 +6,19 @@ import (
 	"github.com/algorand/go-algorand/protocol"
 )
 
-// StateProofMessage represents the message to be certified.
-type StateProofMessage struct {
+// Message represents the message to be certified.
+type Message struct {
 	_struct struct{} `codec:",omitempty,omitemptyarray"`
 	Payload []byte   `codec:"p"`
 }
 
 // ToBeHashed returns the bytes of the message.
-func (m StateProofMessage) ToBeHashed() (protocol.HashID, []byte) {
+func (m Message) ToBeHashed() (protocol.HashID, []byte) {
 	return protocol.CompactCertMessage, m.Payload
 }
 
 // IntoStateProofMessageHash returns a hashed representation fitting the compact certificate messages.
-func (m StateProofMessage) IntoStateProofMessageHash() cc.StateProofMessageHash {
+func (m Message) IntoStateProofMessageHash() cc.StateProofMessageHash {
 	digest := crypto.GenericHashObj(crypto.HashFactory{HashType: cc.StateProofMessageHashType}.NewHash(), m)
 	result := cc.StateProofMessageHash{}
 	copy(result[:], digest)
